Divide js code into modules
