+++ index.html
@@ -7,7 +7,7 @@
   <link rel="preconnect" href="https://fonts.googleapis.com">
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
   <link href="https://fonts.googleapis.com/css2?family=Lobster&display=swap" rel="stylesheet">
-  <link rel="stylesheet" href="style.css">
+  <link rel="stylesheet" href="index.css">
   <title>Awesome books</title>
 </head>
 <body>
@@ -15,9 +15,9 @@
     <nav>
       <a href="#">Awesome books</a>
       <ul class="nav-menu">
-        <li class="nav-items">List</li>
-        <li class="nav-items">Add new</li>
-        <li class="nav-items">Contact</li>
+        <li class="nav-items list-item">List</li>
+        <li class="nav-items add-item">Add new</li>
+        <li class="nav-items contact-item">Contact</li>
       </ul>
     </nav>
   </header>
@@ -65,6 +65,7 @@
   <footer>
     <h3>Copyright</h3>
   </footer>
-  <script src="script.js"></script>
+  <script type="module" src="index.js"></script>
+  <script src="https://moment.github.io/luxon/global/luxon.min.js"></script>
 </body>
 </html>
